removed firebase auth
