+++ src/contexts/AuthContext.js
@@ -1,55 +1,55 @@
-import React, { useContext, useEffect, useState } from "react";
-import { auth } from "../firebase";
+// import React, { useContext, useEffect, useState } from "react";
+// import { auth } from "../firebase";
 
-const AuthContext = React.createContext();
+// const AuthContext = React.createContext();
 
-export function useAuth() {
-  return useContext(AuthContext);
-}
+// export function useAuth() {
+//   return useContext(AuthContext);
+// }
 
-export function AuthProvider({ children }) {
-  const [currentUser, setCurrentUser] = useState();
-  const [loading, setLoading] = useState(true);
+// export function AuthProvider({ children }) {
+//   const [currentUser, setCurrentUser] = useState();
+//   const [loading, setLoading] = useState(true);
 
-  const signup = (email, password) => {
-    return auth.createUserWithEmailAndPassword(email, password);
-  };
-  const signin = (email, password) => {
-    return auth.signInWithEmailAndPassword(email, password);
-  };
-  const signout = () => {
-    return auth.signOut();
-  };
-  const resetPassword = (email) => {
-    return auth.sendPasswordResetEmail(email);
-  };
-  const updateEmail = (email) => {
-    return currentUser.updateEmail(email);
-  };
-  const updatePassword = (password) => {
-    return currentUser.updatePassword(password);
-  };
+//   const signup = (email, password) => {
+//     return auth.createUserWithEmailAndPassword(email, password);
+//   };
+//   const signin = (email, password) => {
+//     return auth.signInWithEmailAndPassword(email, password);
+//   };
+//   const signout = () => {
+//     return auth.signOut();
+//   };
+//   const resetPassword = (email) => {
+//     return auth.sendPasswordResetEmail(email);
+//   };
+//   const updateEmail = (email) => {
+//     return currentUser.updateEmail(email);
+//   };
+//   const updatePassword = (password) => {
+//     return currentUser.updatePassword(password);
+//   };
 
-  useEffect(() => {
-    const unsubscribe = auth.onAuthStateChanged((user) => {
-      setCurrentUser(user);
-      setLoading(false);
-    });
-    return unsubscribe;
-  }, []);
+//   useEffect(() => {
+//     const unsubscribe = auth.onAuthStateChanged((user) => {
+//       setCurrentUser(user);
+//       setLoading(false);
+//     });
+//     return unsubscribe;
+//   }, []);
 
-  const value = {
-    currentUser,
-    signup,
-    signin,
-    signout,
-    resetPassword,
-    updateEmail,
-    updatePassword,
-  };
-  return (
-    <AuthContext.Provider value={value}>
-      {!loading && children}
-    </AuthContext.Provider>
-  );
-}
+//   const value = {
+//     currentUser,
+//     signup,
+//     signin,
+//     signout,
+//     resetPassword,
+//     updateEmail,
+//     updatePassword,
+//   };
+//   return (
+//     <AuthContext.Provider value={value}>
+//       {!loading && children}
+//     </AuthContext.Provider>
+//   );
+// }
